Analysing altruistic behaviour (captured by the ultimatum game) diffusion in a Barabasi-Albert network.
This model shows how altruistic behaviour emerges in social networks
from an entirely selfish group. The analysis is conducted by simulating
the ultimatum game on a Barabasi-Albert network with different degree
of preferential attachment. While previous work studied the emergence
of altruistic behaviour in a well mixed population, this paper will
study how social networks can limit the arise of altruism. Agents in
the model are empathic and memoryless and the strategic update rule is
natural selection.